Remove util module
